requirements.txt
colormath
PySimpleGUI
pyinstaller
Pillow